+++ src/main/java/edu/school21/infoweb/repositories/FriendsRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FriendsRepository extends CrudRepository<Friends, Integer> {
+public interface FriendsRepository extends CrudRepository<Friends, Long> {
     List<Friends> findAll();
 }
